Added two more recipes and updated index.html
This update added a chicken-jerk.html and cookie-chocochip.html both
fully up and working, the recipes taste pretty good, so don't be a afraid
to try them. The update also updated index.html with the links to the
new pages.

Next update:
* More recipes
* Restyling the pages

Update was expected to come out over the weekend but the after effects
of covid got me and I could not look at a computer screen. At least I
got some much needed sleep?.. The next update will probably take a while
, Need to learn CSS and some JavaScript before making the website all
nice. Thanks to anybody taking the time out of their day to read this!
-Munchnax