Updated and fixed Webpages and Navigation
This update was to add a return to home link on the recipes pages and
fix some grammatical and image sizing errors. Also updated the homepage
to make it feel less empty and more home-pagey(?).

This update wasn't really needed as I will probably redo this in CSS but
I figured making it look better, would help more than harm... Also I was
bored. Btw I figured out how to delete suggestions, just delete the
items in the project with the name. Their is probably a better way to
complete this but, it worked for me.
-Munchnax